sensor/requirements.txt
RPi.GPIO
boto3==1.7.31
google-auth==1.5.1
gspread==3.0.0
oauth2client==4.1.2
picamera==1.13
